+++ lib/workable.rb
@@ -3,11 +3,14 @@ require 'uri'
 require 'net/http'
 require 'date'
 require 'ostruct'
+require 'cgi'
 
-require_relative "workable/version"
-require_relative "workable/errors"
-require_relative "workable/client"
+require_relative 'workable/version'
+require_relative 'workable/errors'
+require_relative 'workable/client'
+require_relative 'workable/transformation'
+require_relative 'workable/collection'
 
 module Workable
-  API_VERSION = "2".freeze
+  API_VERSION = 3
 end