+++ lib/workable/client.rb
@@ -1,13 +1,12 @@
 module Workable
   class Client
-
     # set access to workable and data transformation methods
     #
     # @param options [Hash]
     # @option options :api_key   [String] api key for workable
     # @option options :subdomain [String] company subdomain in workable
     # @option options :transform_to [Hash<Symbol: Proc>] mapping of transformations for data
-    #    available transformations: [:job, :candidate, :question, :stage]
+    #    available transformations: [:job, :candidate, :question, :stage, :recruiter, :member]
     #    when no transformation is given raw Hash / Array data is returned
     #
     # @example transformation for candidates using `MyApp::Candidate.find_and_update_or_create`
@@ -29,60 +28,103 @@ module Workable
     #      }
     #    )
     def initialize(options = {})
-      @api_key   = options.fetch(:api_key)   { fail Errors::InvalidConfiguration, "Missing api_key argument"   }
-      @subdomain = options.fetch(:subdomain) { fail Errors::InvalidConfiguration, "Missing subdomain argument" }
-      @transform_to   = options[:transform_to]   || {}
-      @transform_from = options[:transform_from] || {}
+      @api_key   = options.fetch(:api_key)   { configuration_error 'Missing api_key argument'   }
+      @subdomain = options.fetch(:subdomain) { configuration_error 'Missing subdomain argument' }
+      @transform_to   = Transformation.new(options[:transform_to])
+      @transform_from = Transformation.new(options[:transform_from])
     end
 
-    # request jobs of given type
-    # @param type [String] type of jobs to fetch, `published` by default
-    def jobs(type = 'published')
-      transform_to(:job, get_request("jobs?phase=#{type}")['jobs'])
+    # return information about your account
+    def about
+      get_request('')
+    end
+
+    # returns a collection of your account members
+    def members
+      @transform_to.apply(:member, get_request('members')['members'])
+    end
+
+    # returns a collection of your account external recruiters
+    def recruiters
+      @transform_to.apply(:recruiter, get_request('recruiters')['recruiters'])
+    end
+
+    # returns a collection of your recruitment pipeline stages
+    def stages
+      @transform_to.apply(:stage, get_request('stages')['stages'])
+    end
+
+    # request posted jobs
+    # @option params [Hash] optional filter parameters
+    # @option params :state [String] Returns jobs with the current state. Possible values (draft, published, archived & closed)
+    # @option params :limit [Integer] Specifies the number of jobs to try and retrieve per page
+    # @option params :since_id [String] Returns results with an ID more than or equal to the specified ID.
+    # @option params :max_id [String] Returns results with an ID less than or equal to the specified ID.
+    # @option params :created_after [Timestamp|Integer] Returns results created after the specified timestamp.
+    # @option params :updated_after [Timestamp|Integer] Returns results updated after the specified timestamp.
+    def jobs(params = {})
+      build_collection('jobs', :job, 'jobs', params)
     end
 
     # request detailed information about job
     # @param shortcode [String] job short code
     def job_details(shortcode)
-      transform_to(:job, get_request("jobs/#{shortcode}"))
-    end
-
-    # list candidates for given job
-    # @param  shortcode [String] job shortcode to select candidates from
-    # @param  options   [Hash]   extra options like `stage_slug` or `limit`
-    # @option options :stage [String]        optional stage slug, if not given candidates are listed for all stages
-    # @option options :limit [Number|String] optional limit of candidates to download, if not given all candidates are listed
-    def job_candidates(shortcode, options = {})
-      url = build_job_candidates_url(shortcode, options)
-      transform_to(:candidate, get_request(url)['candidates'])
+      @transform_to.apply(:job, get_request("jobs/#{shortcode}"))
     end
 
     # list of questions for job
     # @param shortcode [String] job short code
     def job_questions(shortcode)
-      transform_to(:question, get_request("jobs/#{shortcode}/questions")['questions'])
+      @transform_to.apply(:question, get_request("jobs/#{shortcode}/questions")['questions'])
+    end
+
+    # return a collection of job's members
+    # @param shortcode [String] job short code
+    def job_members(shortcode)
+      @transform_to.apply(:member, get_request("jobs/#{shortcode}/members")['members'])
+    end
+
+    # return a collection of the job's external recruiters
+    # @param shortcode [String] job short code
+    def job_recruiters(shortcode)
+      @transform_to.apply(:recruiter, get_request("jobs/#{shortcode}/recruiters")['recruiters'])
+    end
+
+    # list candidates for given job
+    # @param  shortcode [String] job shortcode to select candidates from
+    # @param  params [Hash]   extra options like `state` or `limit`
+    # @option params :state [String]        optional state slug, if not given candidates are listed for all stages
+    # @option params :limit [Number|String] optional limit of candidates to download, if not given all candidates are listed
+    # @option params :since_id [String] Returns results with an ID more than or equal to the specified ID.
+    # @option params :max_id [String] Returns results with an ID less than or equal to the specified ID.
+    # @option params :created_after [Timestamp|Integer] Returns results created after the specified timestamp.
+    # @option params :updated_after [Timestamp|Integer] Returns results updated after the specified timestamp.
+    def job_candidates(shortcode, params = {})
+      build_collection("jobs/#{shortcode}/candidates", :candidate, 'candidates', params)
+    end
+
+    # return the full object of a specific candidate
+    # @param shortcode [String] job shortcode to select candidate from
+    # @param id [String] candidates's id
+    def job_candidate(shortcode, id)
+      @transform_to.apply(:candidate, get_request("jobs/#{shortcode}/candidates/#{id}")['candidate'])
     end
 
     # create new candidate for given job
     # @param candidate  [Hash] the candidate data as described in
-    #    http://resources.workable.com/add-candidates-using-api
+    #    https://workable.readme.io/docs/job-candidates-create
     #    including the `{"candidate"=>{}}` part
     # @param shortcode  [String] job short code
     # @param stage_slug [String] optional stage slug
     # @return [Hash] the candidate information without `{"candidate"=>{}}` part
     def create_job_candidate(candidate, shortcode, stage_slug = nil)
-      shortcode = "#{shortcode}/#{stage_slug}" unless stage_slug.nil?
-      transform_to(:candidate, post_request("jobs/#{shortcode}/candidates", candidate)["candidate"])
-    end
+      shortcode = "#{shortcode}/#{stage_slug}" if stage_slug
 
-    # list of stages defined for company
-    def stages
-      transform_to(:stage, get_request("stages")['stages'])
-    end
+      response = post_request("jobs/#{shortcode}/candidates") do |request|
+        request.body = @transform_from.apply(:candidate, candidate).to_json
+      end
 
-    # list of external recruiters for company
-    def recruiters
-      transform_to(:stage, get_request("recruiters")['recruiters'])
+      @transform_to.apply(:candidate, response['candidate'])
     end
 
     private
@@ -91,28 +133,30 @@ module Workable
 
     # build the url to api
     def api_url
-      "https://www.workable.com/spi/v%s/accounts/%s" % [Workable::API_VERSION, subdomain]
+      'https://www.workable.com/spi/v%s/accounts/%s' % [Workable::API_VERSION, subdomain]
     end
 
     # do the get request to api
-    def get_request(url)
-      do_request(url, Net::HTTP::Get)
+    def get_request(url, params = {})
+      params = URI.encode_www_form(params.keep_if { |k, v| k && v })
+      full_url = params.empty? ? url : [url, params].join('?')
+      do_request(full_url, Net::HTTP::Get)
     end
 
     # do the post request to api
-    def post_request(url, data)
+    def post_request(url)
       do_request(url, Net::HTTP::Post) do |request|
-        request.body = transform_from(:candidate, data).to_json
+        yield(request) if block_given?
       end
     end
 
     # generic part of requesting api
-    def do_request(url, type, &block)
+    def do_request(url, type, &_block)
       uri = URI.parse("#{api_url}/#{url}")
       http = Net::HTTP.new(uri.host, uri.port)
       http.use_ssl = true
 
-      request  = type.new(uri.request_uri, headers)
+      request = type.new(uri.request_uri, headers)
       yield request if block_given?
       response = http.request(request)
 
@@ -127,28 +171,26 @@ module Workable
       when 200...300 # handled with response
         JSON.parse(response.body)
       when 401
-        raise Errors::NotAuthorized, JSON.parse(response.body)["error"]
+        fail Errors::NotAuthorized, JSON.parse(response.body)['error']
       when 404
-        raise Errors::NotFound, JSON.parse(response.body)["error"]
+        fail Errors::NotFound, JSON.parse(response.body)['error']
       when 422
         handle_response_422(response)
       when 503
-        raise Errors::RequestToLong, response.body
+        fail Errors::RequestToLong, response.body
       else
-        raise Errors::InvalidResponse, "Response code: #{response.code} message: #{response.body}"
+        fail Errors::InvalidResponse, "Response code: #{response.code} message: #{response.body}"
       end
     end
 
     def handle_response_422(response)
       data = JSON.parse(response.body)
-      if
-        data["validation_errors"] &&
-        data["validation_errors"]["email"] &&
-        data["validation_errors"]["email"].include?("candidate already exists")
-      then
-        raise Errors::AlreadyExists, data["error"]
+      if data['validation_errors'] &&
+         data['validation_errors']['email'] &&
+         data['validation_errors']['email'].include?('candidate already exists')
+        fail Errors::AlreadyExists, data['error']
       else
-        raise Errors::NotFound, data["error"]
+        fail Errors::NotFound, data['error']
       end
     end
 
@@ -158,33 +200,8 @@ module Workable
         'Accept'        => 'application/json',
         'Authorization' => "Bearer #{api_key}",
         'Content-Type'  => 'application/json',
-        'User-Agent'    => 'Workable Ruby Client',
+        'User-Agent'    => 'Workable Ruby Client'
       }
-    end
-
-    # build url for fetching job candidates
-    # @param  shortcode [String] job shortcode to select candidates from
-    # @param  options   [Hash]   extra options like `stage_slug` or `limit`
-    # @option options :stage_slug [String]        optional stage slug, if not given candidates are listed for all stages
-    # @option options :limit      [Number|String] optional limit of candidates to download, if not given all candidates are listed
-    def build_job_candidates_url(shortcode, options)
-      if (stage_slug = options.delete(:stage))
-      then stage_slug = "/#{stage_slug}"
-      end
-      params =
-      if options.empty?
-      then ""
-      else "?#{options.map{|k,v| "#{k}=#{v}"}.join("&")}"
-      end
-      "jobs/#{shortcode}#{stage_slug}/candidates#{params}"
-    end
-
-    # transform result using given method if defined
-    # @param type [Symbol] type of the transformation, one of `[:job, :candidate, :question, :stage]`
-    # @param result [Hash|Array|nil] the value to transform, can be nothing, `Hash` of values or `Array` of `Hash`es
-    # @return transformed result if transformation exists for type, raw result otherwise
-    def transform_to(type, result)
-      transform(@transform_to[type], result)
     end
 
     # transform input using given method if defined
@@ -195,22 +212,20 @@ module Workable
       transform(@transform_from[type], input)
     end
 
-    # selects transformation strategy based on the inputs
-    # @param transformation [Method|Proc|nil] the transformation to perform
-    # @param data           [Hash|Array|nil]  the data to transform
-    # @return               [Object|nil]
-    #    results of the transformation if given, raw data otherwise
-    def transform(transformation, data)
-      return data unless transformation
-      case data
-      when nil
-        data
-      when Array
-        data.map{|datas| transformation.call(datas) }
-      else
-        transformation.call(data)
-      end
+    def build_collection(url, transform_mapping, root_key, params = {})
+      url = url.gsub(/#{api_url}\/?/, '')
+      response = get_request(url, params)
+
+      Collection.new(
+        @transform_to.apply(transform_mapping, response[root_key]),
+        method(__callee__),
+        transform_mapping,
+        root_key,
+        response['paging'])
     end
 
+    def configuration_error(message)
+      fail Errors::InvalidConfiguration, message
+    end
   end
 end