+++ spec/spec_helper.rb
@@ -1,5 +1,10 @@
+require 'simplecov'
 require 'coveralls'
-Coveralls.wear!
+
+SimpleCov.formatter = Coveralls::SimpleCov::Formatter
+SimpleCov.start do
+  add_filter 'spec/'
+end
 
 require_relative '../lib/workable'
 require_relative 'fixtures'